ml_model/requirements.txt
pandas==2.0.3
numpy==1.24.3
scikit-learn==1.2.2
tensorflow==2.12.0
flask==2.2.2
python-dotenv==1.0.0
joblib==1.2.0
